hackuta-backend/requirements.txt
# basic backend 
fastapi[standard]
fastapi-cli
uvicorn[standard]
pydantic
python-dotenv
python-multipart

# database - using aiosqlite for local dev (no separate DB server needed)
# For production PostgreSQL, also install: asyncpg
sqlalchemy
aiosqlite
greenlet

# s3 sdk
boto3

# http requests
requests
aiohttp

# JWT authentication
python-jose[cryptography]
python-multipart

# ML stuff (for future use)
# pillow==10.1.0
# transformers==4.36.0
# torch==2.1.0
# openai==1.3.0
# google-generativeai==0.3.0
